tss/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-import-export==2.8.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fontawesomefree==6.1.1
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.9.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
phonenumbers==8.12.46
Pillow==9.0.1
pkg_resources==0.0.0
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
style==1.1.0
svglib==1.3.0
tablib==3.2.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
update==0.0.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
